Moved most of the implementation of the methods of the LoopNest class to LoopNest.cpp
