+++ apps/support/autoscheduler.inc
@@ -54,6 +54,8 @@ $(AUTOSCHED_BIN)/libauto_schedule.so: $(AUTOSCHED_SRC)/AutoSchedule.cpp \
 										$(AUTOSCHED_SRC)/Weights.cpp \
 										$(AUTOSCHED_SRC)/FunctionDAG.h \
 										$(AUTOSCHED_SRC)/FunctionDAG.cpp \
+										$(AUTOSCHED_SRC)/LoopNest.h \
+										$(AUTOSCHED_SRC)/LoopNest.cpp \
 										$(AUTOSCHED_SRC)/Featurization.h \
 										$(AUTOSCHED_SRC)/CostModel.h \
 										$(AUTOSCHED_SRC)/PerfectHashMap.h \
